MRVoxels: revert the MRFloatGrid / MRVDBConversions TU splits
The splits (folded in from #6286/#6288) parallelized the OpenVDB codegen by
duplicating it across the new TUs -- Build Insights showed they tripled the
codegen (80 -> 265 s) for zero wall benefit. The extern-template instantiation,
which homes that codegen once, is the correct fix and makes the splits
redundant. Keep PCH + extern-template; restore MRFloatGrid.cpp and
MRVDBConversions.cpp to single TUs and drop the 5 split files.
